move types to ./types
