+++ src/services/l1-ingestion/handlers/sequencer-batch-appended.ts
@@ -13,6 +13,8 @@ import {
 import {
   DecodedSequencerBatchTransaction,
   EventArgsSequencerBatchAppended,
+  SequencerBatchAppendedExtraData,
+  SequencerBatchAppendedParsedEvent,
   TransactionBatchEntry,
   TransactionEntry,
   EventHandlerSet,
@@ -21,27 +23,6 @@ import {
   SEQUENCER_ENTRYPOINT_ADDRESS,
   SEQUENCER_GAS_LIMIT,
 } from '../../../utils'
-
-export interface SequencerBatchAppendedExtraData {
-  timestamp: number
-  blockNumber: number
-  submitter: string
-  l1TransactionData: string
-  l1TransactionHash: string
-  gasLimit: number
-
-  // Stuff from TransactionBatchAppended.
-  prevTotalElements: BigNumber
-  batchIndex: BigNumber
-  batchSize: BigNumber
-  batchRoot: string
-  batchExtraData: string
-}
-
-export interface SequencerBatchAppendedParsedEvent {
-  transactionBatchEntry: TransactionBatchEntry
-  transactionEntries: TransactionEntry[]
-}
 
 export const handleEventsSequencerBatchAppended: EventHandlerSet<
   EventArgsSequencerBatchAppended,
