+++ test/unit-tests/services/l1-ingestion/handlers/state-batch-appended.spec.ts
@@ -2,10 +2,8 @@ import { BigNumber } from 'ethers'
 import { Block } from '@ethersproject/abstract-provider'
 
 import { expect } from '../../../../setup'
-import {
-  handleEventsStateBatchAppended,
-  StateBatchAppendedExtraData,
-} from '../../../../../src/services/l1-ingestion/handlers/state-batch-appended'
+import { handleEventsStateBatchAppended } from '../../../../../src/services/l1-ingestion/handlers/state-batch-appended'
+import { StateBatchAppendedExtraData } from '../../../../../src/types'
 import { l1StateBatchData } from '../../../examples/l1-data'
 
 describe('Event Handlers: OVM_CanonicalTransactionChain.StateBatchAppended', () => {
